requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
dlib==19.4.0
docopt==0.6.2
efficientnet==1.1.0
Flask==1.1.1
gast==0.2.2
gevent==1.4.0
google-images-download==2.8.0
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
hyperlink==19.0.0
idna==2.8
imageio==2.8.0
imgaug==0.4.0
incremental==17.5.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.12.1
Jinja2==2.10.3
joblib==0.14.1
jupyter-client==5.3.3
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lmdb==0.98
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.3
networkx==2.4
numpy==1.16.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
parsel==1.5.2
parso==0.5.2
pickleshare==0.7.4
Pillow==7.0.0
pipreqs==0.4.10
prompt-toolkit==1.0.15
Protego==0.1.16
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybloom==2.0.0
pycparser==2.19
PyDispatcher==2.0.5
PyFaceDet==0.2.0
Pygments==2.5.2
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==17.1.2
queuelib==1.5.0
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.2.1
Scrapy==2.0.0
selenium==3.141.0
service-identity==18.1.0
Shapely==1.7.0
simplegeneric==0.8.1
six==1.13.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==5.1.1
tqdm==4.43.0
traitlets==4.3.2
Twisted==19.10.0
urllib3==1.24.2
w3lib==1.19.0
wcwidth==0.1.7
Werkzeug==0.16.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
yarg==0.1.9
zope.interface==4.7.2
